requirements.txt
fastapi
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic
